Add: folders to separate config data for each client page
